// embedded-components/src/api/generated/smbdo.schemas.ts
export interface AttestationRemoval {
  documentId: DocumentId;
}

/**
 * Update a party inline within the clients API by supplying its ID.
 */
export interface UpdatePartyRequestInline {
  access?: AccessList;
  email?: Email;
  id?: PartyId;
  individualDetails?: IndividualDetails;
  organizationDetails?: OrganizationDetails;
  /**
   * @minItems 1
   * @maxItems 10
   */
  roles?: Role[];
}

export type UpdateClientRequestSmbdoAddPartiesItem = CreatePartyRequestInline &
  UpdatePartyRequestInline;

export interface UpdateClientRequestSmbdo {
  /**
   * @minItems 1
   * @maxItems 10
   */
  addAttestations?: Attestation[];
  /**
   * Update a party by specifying its ID in the object. Create a party by setting `parentPartyId` to the client's root party. `partyType` and `roles` are required when creating a new party.

   * @minLength 1
   * @maxLength 10
   * @minItems 0
   */
  addParties?: UpdateClientRequestSmbdoAddPartiesItem[];
  addProducts?: ClientProductList;
  consumerDevice?: ConsumerDevice;
  /**
   * @minItems 0
   * @maxItems 200
   */
  questionResponses?: ClientQuestionResponse[];
  /**
   * @minItems 1
   * @maxItems 10
   */
  removeAttestations?: AttestationRemoval[];
}

export type ClientUpdatedResponse = ClientResponse & {
  consumerDevice?: ConsumerDevice;
};

export type ClientResponseOutstanding = {
  attestationDocumentIds?: DocumentIdList;
  documentRequestIds?: DocumentRequestIdList;
  partyIds?: PartyIdList;
  partyRoles?: PartyRoleList;
  questionIds?: QuestionIdList;
};

/**
 * The status of the customer identification process.
 */
export type CustomerIdentityStatus =
  (typeof CustomerIdentityStatus)[keyof typeof CustomerIdentityStatus];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const CustomerIdentityStatus = {
  APPROVED: 'APPROVED',
  INFORMATION_REQUESTED: 'INFORMATION_REQUESTED',
  NOT_STARTED: 'NOT_STARTED',
  REVIEW_IN_PROGRESS: 'REVIEW_IN_PROGRESS',
} as const;

/**
 * The results of the client verification process.
 */
export interface ClientResults {
  customerIdentityStatus: CustomerIdentityStatus;
}

/**
 * The values of a question response. Each item is a string that will be validated according to the question's responseSchema.

 * @minItems 1
 * @maxItems 20
 */
export type ResponseValueList = string[];

/**
 * The client's response to a question.
 */
export interface ClientQuestionResponse {
  questionId?: QuestionId;
  values?: ResponseValueList;
}

export interface PartyResponse {
  /** @deprecated */
  access?: AccessList;
  active?: Active;
  createdAt?: string;
  email?: Email;
  externalId?: ExternalId;
  id?: PartyId;
  individualDetails?: IndividualDetails;
  networkRegistration?: NetworkRegistration;
  organizationDetails?: OrganizationDetails;
  parentExternalId?: ExternalId;
  parentPartyId?: ParentPartyId;
  partyType?: PartyType;
  profileStatus?: ProfileStatus;
  /**
   * @minItems 0
   * @maxItems 10
   */
  roles?: Role[];
  /** @deprecated */
  status?: PartyStatus;
  validationResponse?: ValidationResponse;
}

export interface ClientResponse {
  /**
   * @minItems 0
   * @maxItems 10
   */
  attestations?: Attestation[];
  /** Date and time when the client was created */
  createdAt?: string;
  id: ClientId;
  outstanding: ClientResponseOutstanding;
  /** @minItems 0 */
  parties?: PartyResponse[];
  partyId: PartyId;
  products: ClientProductList;
  /**
   * @minItems 0
   * @maxItems 200
   */
  questionResponses?: ClientQuestionResponse[];
  results?: ClientResults;
  status: ClientStatus;
}
// ...
export interface NetworkRegistration {
  amexOptBlueMarketing?: AmexOptBlueMarketing;
  amexSellerId?: AmexSellerId;
  currencyCode?: NetworkRegistrationCurrencyCode;
  discoverDebtRepayment?: DiscoverDebtRepayment;
  discoverMerchantId?: DiscoverMerchantId;
  jcbAccepted?: JcbAccepted;
  serviceEstablishmentStatus?: ServiceEstablishmentStatus;
  /** The last update date of the service establishment status.  Required if the service establishment status is provided.
   */
  serviceEstablishmentStatusUpdateDate?: string;
  /**
   * The selection of method of payments used in network registration.
   * @minItems 1
   * @maxItems 3
   */
  settlementPaymentMethods?: SettlementPaymentMethods[];
  /** The opting out of settlement payment method selection. Required to specify true if the merchant is not selecting any settlement payment method.
   */
  settlementPaymentMethodsOptOut?: boolean;
  transactionDeviceType?: TransactionDeviceType;
  visaDebitAcceptance?: VisaDebitAcceptance;
}

/**
 * List of validation info.
 * @minItems 0
 * @maxItems 10
 */
export type ValidationResponse = ValidationResponseItem[];
// ...
export type Active = boolean;

export type PartyStatus = (typeof PartyStatus)[keyof typeof PartyStatus];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const PartyStatus = {
  ACTIVE: 'ACTIVE',
  INACTIVE: 'INACTIVE',
} as const;

export type ProfileStatus = (typeof ProfileStatus)[keyof typeof ProfileStatus];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ProfileStatus = {
  NEW: 'NEW',
  REVIEW_IN_PROGRESS: 'REVIEW_IN_PROGRESS',
  INFORMATION_REQUESTED: 'INFORMATION_REQUESTED',
  APPROVED: 'APPROVED',
  DECLINED: 'DECLINED',
  SUSPENDED: 'SUSPENDED',
  TERMINATED: 'TERMINATED',
} as const;

/**
 * The party type
 */
export type PartyType = (typeof PartyType)[keyof typeof PartyType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const PartyType = {
  INDIVIDUAL: 'INDIVIDUAL',
  ORGANIZATION: 'ORGANIZATION',
} as const;

export type OrganizationDetailsRequired = OrganizationDetails;

/**
 * Create a party within the clients payload.
 */
export interface CreatePartyRequestInline {
  access?: AccessList;
  email?: Email;
  externalId?: ExternalId;
  individualDetails?: IndividualDetailsRequired;
  organizationDetails?: OrganizationDetailsRequired;
  parentPartyId?: ParentPartyId;
  partyType?: PartyTypeSmbdo;
  /**
   * @minItems 1
   * @maxItems 10
   */
  roles?: Role[];
}

/**
 * Create a party within the clients payload.
 */
export type CreatePartyRequestInlineRequired = CreatePartyRequestInline;

export interface CreateClientRequestSmbdo {
  /**
   * @minItems 1
   * @maxItems 10
   */
  attestations?: Attestation[];
  consumerDevice?: ConsumerDevice;
  /**
   * @minItems 1
   * @maxItems 10
   */
  parties?: CreatePartyRequestInlineRequired[];
  partyId?: PartyId;
  products: ClientProductList;
}
// ...
export interface OrganizationDetails {
  /**
   * @minItems 1
   * @maxItems 5
   */
  addresses?: AddressDto[];
  /**
   * @minItems 0
   * @maxItems 100
   */
  associatedCountries?: CountryCodeIsoAlpha2[];
  countryOfFormation?: CountryCodeIsoAlpha2;
  dbaName?: DbaName;
  entitiesInOwnership?: EntitiesInOwnership;
  industry?: OrganizationIndustry;
  industryCategory?: OrganizationIndustryCategory;
  industryType?: OrganizationIndustryType;
  jurisdiction?: CountryCodeIsoAlpha2;
  mcc?: MerchantCategoryCode;
  organizationDescription?: OrganizationDescription;
  /**
   * @minItems 0
   * @maxItems 6
   */
  organizationIds?: OrganizationIdentityDto[];
  organizationName?: OrganizationName;
  organizationType?: OrganizationType;
  phone?: PhoneSmbdo;
  /**
   * The list of additional merchant category codes describing  industries that the business is in.

   * @minItems 0
   * @maxItems 10
   */
  secondaryMccList?: MerchantCategoryCode[];
  socialMedia?: SocialMediaList;
  website?: Website;
  websiteAvailable?: WebsiteAvailable;
  yearOfFormation?: YearOfFormation;
}
// ...
export interface IndividualDetails {
  /**
   * @minItems 1
   * @maxItems 5
   */
  addresses?: AddressDto[];
  birthDate?: BirthDate;
  countryOfResidence?: CountryCodeIsoAlpha2;
  firstName?: FirstName;
  /**
   * @minItems 0
   * @maxItems 16
   */
  individualIds?: IndividualIdentity[];
  jobTitle?: IndividualJobTitle;
  jobTitleDescription?: IndividualJobTitleDescription;
  lastName?: LastName;
  middleName?: MiddleName;
  nameSuffix?: NameSuffix;
  natureOfOwnership?: NatureOfOwnership;
  phone?: PhoneSmbdo;
  socialMedia?: SocialMediaList;
  soleOwner?: SoleOwner;
}
// ...
export type ParentPartyId = string;

/**
 * Id in external system.
 * @minLength 1
 * @maxLength 20
 */
export type ExternalId = string;

/**
 * Email of the party.
 * @minLength 0
 * @maxLength 320
 */
export type Email = string;

/**
 * List of user access descriptors.
 * @deprecated
 * @minItems 0
 * @maxItems 10
 */
export type AccessList = string[];

/**
 * Consumer device information provided by the merchant. It is only returned when present in the create or update client requests.
 */
export interface ConsumerDevice {
  /** IPv4 address of the consumer device. */
  ipAddress?: string;
  /**
   * The unique session identifier of the device that was created by the client. It can be up to 128 characters long and can contain only the following characters: uppercase and lowercase Roman letters, digits, underscore characters, and hyphens (a?z, A?Z, 0?9, _, -). The session ID should contain at least 16 bytes of randomly generated data.

   * @pattern ^[a-zA-Z0-9_-]+$
   */
  sessionId?: string;
}

/**
 * The attesting of presented legal statements. Either "attester" or "attesterFullName" is required.
 */
export interface Attestation {
  attestationTime: string;
  attester?: Attester;
  /** The full name of an individual. */
  attesterFullName?: string;
  documentId: DocumentId;
  /** IPv4 Address. */
  ipAddress: string;
}
// ...
export interface Attester {
  designation: AttesterDesignation;
  firstName: AttesterFirstName;
  lastName: AttesterLastName;
  middleName?: AttesterMiddleName;
}
// ...
export type ClientStatus = (typeof ClientStatus)[keyof typeof ClientStatus];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ClientStatus = {
  APPROVED: 'APPROVED',
  DECLINED: 'DECLINED',
  INFORMATION_REQUESTED: 'INFORMATION_REQUESTED',
  NEW: 'NEW',
  REVIEW_IN_PROGRESS: 'REVIEW_IN_PROGRESS',
  SUSPENDED: 'SUSPENDED',
  TERMINATED: 'TERMINATED',
} as const;

export interface ClientSummaryResponse {
  id: ClientId;
  outstanding: ClientSummaryResponseOutstanding;
  partyId: PartyId;
  products: ClientProductList;
  status: ClientStatus;
}

/**
 * Collection of clients.
 */
export interface ClientListResponse {
  /**
   * @minItems 0
   * @maxItems 100
   */
  clients: ClientSummaryResponse[];
  metadata: PageMetaData;
}

export type Role = (typeof Role)[keyof typeof Role];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const Role = {
  AUTHORIZED_USER: 'AUTHORIZED_USER',
  BENEFICIAL_OWNER: 'BENEFICIAL_OWNER',
  CLIENT: 'CLIENT',
  CONTROLLER: 'CONTROLLER',
  DECISION_MAKER: 'DECISION_MAKER',
  PRIMARY_CONTACT: 'PRIMARY_CONTACT',
  DIRECTOR: 'DIRECTOR',
} as const;

/**
 * A list of party roles.
 * @minItems 0
 * @maxItems 10
 */
export type PartyRoleList = Role[];

/**
 * The ID uniquely identifying a question.
 * @maxLength 10
 */
export type QuestionId = string;

/**
 * A list of question IDs.
 * @minItems 0
 * @maxItems 200
 */
export type QuestionIdList = QuestionId[];

/**
 * A list of document request IDs.
 * @minItems 0
 * @maxItems 10
 */
export type DocumentRequestIdList = string[];

/**
 * The unique id generated by the system for the uploaded document,  which can be used for future retrieval.

 */
export type DocumentId = string;

/**
 * A list of document IDs.
 * @minItems 0
 * @maxItems 10
 */
export type DocumentIdList = DocumentId[];
// ...
export type ClientProduct = (typeof ClientProduct)[keyof typeof ClientProduct];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ClientProduct = {
  MERCHANT_SERVICES: 'MERCHANT_SERVICES',
  EMBEDDED_PAYMENTS: 'EMBEDDED_PAYMENTS',
} as const;

/**
 * A list of product selections.
 * @minItems 1
 * @maxItems 50
 */
export type ClientProductList = ClientProduct[];

/**
 * Organization/Individual party id, that uniquely identifies the party.

 * @maxLength 10
 * @pattern ^[0-9]{10}$
 */
export type PartyId = string;

/**
 * A list of party IDs.
 * @minItems 0
 * @maxItems 10
 */
export type PartyIdList = PartyId[];

/**
 * A client's unique ID
 * @maxLength 10
 */
export type ClientId = string;
// ...
export type ApiErrorReasonV2Location =
  (typeof ApiErrorReasonV2Location)[keyof typeof ApiErrorReasonV2Location];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ApiErrorReasonV2Location = {
  BODY: 'BODY',
  QUERY: 'QUERY',
  PATH: 'PATH',
  HEADER: 'HEADER',
} as const;

export interface ApiErrorReasonV2 {
  /** Short code that identifies the error - publicly cataloged and documented */
  code?: string;
  /** The location of the property or parameter in error */
  field?: string;
  /** Part of the request which is responsible for the reason */
  location?: ApiErrorReasonV2Location;
  /** Message describing the reason. This message can typically be displayed to your platform's users, except in cases specified otherwise */
  message: string;
}